fix: address review feedback — update docs, move tests to separate file
- Update docs/architecture.md to reflect empty content for tool-data
  messages (clients now render tool_data directly)
- Move tool_data TUI tests to messages_tests.rs per project convention
  (separate test files, not inline mod tests blocks)
