+++ docs/architecture.md
@@ -646,7 +646,7 @@ StreamEvent (NDJSON drain) → extract_assistant_text() → aggregated text
 
 - **`auto_output` flag**: `Message`, `Effect::PostToChannel`, and `ChannelMessageData` carry an `auto_output: bool` field. Only `stream.rs` (`process_lead_output()` and `process_coworker_output()`) sets it to `true` — all other code paths (explicit `midtown channel post`, system messages, nudges) default to `false`. The web UI uses this to apply muted styling (dimmed text + left border) to streamed output, creating visual hierarchy between intentional posts and background output.
 
-- **`process_coworker_output()`** (`daemon/stream.rs`): Takes the set of active coworker session names (excluding the main lead, channel leads, and fork-bound sessions) and posts each coworker's aggregated text output to `dm-<name>`. This includes reviewer sessions — their output streams to DM channels alongside regular coworkers. For messages containing tool calls, the effect carries a lightweight text summary in `content` (e.g., `"[Bash, Read]"`) for CLI TUI visibility, plus structured `tool_data: Vec<ToolBlock>` and `provider: String` for client-side tool-specific rendering in the web UI.
+- **`process_coworker_output()`** (`daemon/stream.rs`): Takes the set of active coworker session names (excluding the main lead, channel leads, and fork-bound sessions) and posts each coworker's aggregated text output to `dm-<name>`. This includes reviewer sessions — their output streams to DM channels alongside regular coworkers. For messages containing tool calls, the effect carries empty `content` with structured `tool_data: Vec<ToolBlock>` and `provider: String` — each client renders tool data its own way (the TUI generates a `[Bash, Read]` text summary locally, the web app renders rich expandable blocks).
 - **Structured tool data** (`ToolBlock` in `message.rs`): Preserves raw tool call JSON (`tool_name`, `input`, `output`, `error`) extracted from stream events. `extract_tool_blocks()` pairs `tool_use` blocks from Assistant events with `tool_result` blocks from User events by `call_id`. `detect_provider()` identifies the AI provider (`"claude"` or `"codex"`) from stream event metadata. Both fields are `Option` with `serde(default)` for backward compatibility with legacy messages.
 - **Nudge content**: When a coworker receives a nudge (task assignment, mention, review, etc.), the nudge message is also posted to `dm-<name>` via `Effect::PostToChannel`. This makes nudge conversations visible in the DM channel alongside coworker output. `DmFromUser` nudges are excluded because the user's message is already written to the DM channel by the RPC post handler before the nudge effect fires. Fork sessions are also excluded — only pool coworker names receive DM posts.
 - **Task separator**: A `PostSystemMessage` separator is posted to `dm-<name>` to visually delineate task boundaries. For regular coworkers this uses the format "─── Task !42: Fix auth bug ───" and happens in three paths: (1) `SpawnSession` when a session spawns, (2) `AssignAndSpawn` when dispatching a new task, and (3) `task.claim` RPC when a coworker self-claims a pending task. Session recovery via `SpawnSession` with `resume=true` omits the separator since one was already posted on initial assignment. Reviewer sessions receive a PR-based separator (e.g., "─── Reviewing PR #42 ───") in their `SpawnCoworkerWithCallbacks` `on_success` effects.
