+++ src/bin/midtown/cli/chat/ui/messages.rs
@@ -1561,74 +1561,6 @@ mod tests {
     }
 
     #[test]
-    fn test_render_message_tool_data_single_tool() {
-        // When content is empty and tool_data has blocks, render_message should
-        // generate a "[ToolName]" summary.
-        let mut msg = test_message("");
-        msg.tool_data = Some(vec![midtown::ToolBlock {
-            tool_name: "Bash".to_string(),
-            input: serde_json::json!({"command": "ls"}),
-            output: None,
-            error: false,
-            call_id: None,
-            parent_tool_use_id: None,
-        }]);
-
-        let tasks = HashMap::new();
-        let lines = render_message(&msg, 80, None, &tasks, None, &[]);
-
-        let all_text: String = lines
-            .iter()
-            .flat_map(|l| l.spans.iter())
-            .map(|s| s.content.as_ref())
-            .collect();
-        assert!(
-            all_text.contains("[Bash]"),
-            "Should render tool summary '[Bash]', got: {}",
-            all_text
-        );
-    }
-
-    #[test]
-    fn test_render_message_tool_data_multiple_tools() {
-        // When content is empty and tool_data has multiple blocks, render_message
-        // should generate a "[Bash, Read]" summary.
-        let mut msg = test_message("");
-        msg.tool_data = Some(vec![
-            midtown::ToolBlock {
-                tool_name: "Bash".to_string(),
-                input: serde_json::json!({"command": "ls"}),
-                output: None,
-                error: false,
-                call_id: None,
-                parent_tool_use_id: None,
-            },
-            midtown::ToolBlock {
-                tool_name: "Read".to_string(),
-                input: serde_json::json!({"file_path": "/tmp/foo"}),
-                output: None,
-                error: false,
-                call_id: None,
-                parent_tool_use_id: None,
-            },
-        ]);
-
-        let tasks = HashMap::new();
-        let lines = render_message(&msg, 80, None, &tasks, None, &[]);
-
-        let all_text: String = lines
-            .iter()
-            .flat_map(|l| l.spans.iter())
-            .map(|s| s.content.as_ref())
-            .collect();
-        assert!(
-            all_text.contains("[Bash, Read]"),
-            "Should render tool summary '[Bash, Read]', got: {}",
-            all_text
-        );
-    }
-
-    #[test]
     fn test_render_content_lines_processes_markdown() {
         // Verifies that render_content_lines applies markdown formatting rather than
         // passing raw syntax through — this is what the thread header now uses.
@@ -1671,3 +1603,7 @@ mod tests {
         );
     }
 }
+
+#[path = "messages_tests.rs"]
+#[cfg(test)]
+mod tool_data_tests;
